modified dtos, organized dtos classes
