--- aiKart/Dtos/AddCardDto.cs
@@ -1,2 +1,0 @@
-namespace aiKart.Dtos;
-public record AddCardDto(string Question, string Answer);
