--- aiKart/Dtos/AddDeckDto.cs
@@ -1,2 +1,0 @@
-namespace aiKart.Dtos;
-public record AddDeckDto(string Name, string Description);
