+++ aiKart/Dtos/CardDtos/AddCardDto.cs
@@ -1,0 +1,2 @@
+namespace aiKart.Dtos.CardDtos;
+public record AddCardDto(int DeckId, string Question, string Answer);
