+++ aiKart/Dtos/CardDtos/CardDto.cs
@@ -1,0 +1,2 @@
+namespace aiKart.Dtos.CardDtos;
+public record CardDto(int Id, int DeckId, string Question, string Answer, string State);
