+++ aiKart/Dtos/CardDtos/CardStateDto.cs
@@ -1,0 +1,2 @@
+namespace aiKart.Dtos.CardDtos;
+public record CardStateDto(string State);
