+++ aiKart/Dtos/CardDtos/UpdateCardDto.cs
@@ -1,0 +1,2 @@
+namespace aiKart.Dtos.CardDtos;
+public record UpdateCardDto(string Question, string Answer);
