+++ aiKart/Dtos/DeckDtos/AddDeckDto.cs
@@ -1,0 +1,2 @@
+namespace aiKart.Dtos.DeckDtos;
+public record AddDeckDto(string Name, string? Description, string? CreatorName);
