+++ aiKart/Dtos/DeckDtos/UpdateDeckDto.cs
@@ -1,0 +1,2 @@
+namespace aiKart.Dtos.DeckDtos;
+public record UpdateDeckDto(string Name, string? Description, string? CreatorName);
